HipDynMem: Prepare for lack of OCL kernel metadata
Skip kernel argument metadata update if the function does not have
OpenCL kernel metadata to begin with (otherwise, we crash). So far the
metadata is optional but level-zero API is known to crash if the
metadata is present but lacks entries for additional kernel arguments.

To-be-upstreamed HIP-clang will omit the OpenCL kernel metadata
because it is not needed by HIPLZ (nor by HIPCL). Also, the metadata
for the address qualifiers was incorrect. E.g. global pointer
arguments had metadata stating them having private address qualifier.